backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
asgiref==3.11.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
Django==5.2.12
django-cors-headers==4.9.0
django-sslserver==0.22
djangorestframework==3.17.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
et_xmlfile==2.0.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
gspread==6.2.1
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.4.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
preshed==3.0.12
proto-plus==1.27.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pysnmp==7.1.22
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32==311
RapidFuzz==3.14.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
thinc==8.3.10
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.2
